+++ packages/resume/lib/Polysemy/Resume/Data/Stop.hs
@@ -1,3 +1,4 @@
+-- |Description: Internal
 module Polysemy.Resume.Data.Stop where
 
 import Polysemy (makeSem)