+++ packages/resume/changelog.md
@@ -1,2 +1,5 @@
+# Unreleased
+* Add combinators `resumeOr` and `resumingOr`, which take an additional branch for the success case.
+
 # 0.1.0.0
 * initial release